Extract markdown editor utils to a separate file
