+++ frontend/static/js/common/utils.js
@@ -1,5 +1,3 @@
-import EasyMDE from "easymde";
-
 export const findParentForm = (element) => {
     let form = element.parentElement;
 
@@ -41,91 +39,6 @@ export const throttle = (fn, wait) => {
         }
     };
 };
-
-const defaultMarkdownEditorOptions = {
-    autoDownloadFontAwesome: false,
-    spellChecker: false,
-    nativeSpellcheck: true,
-    forceSync: true,
-    status: false,
-    inputStyle: "contenteditable",
-    tabSize: 4,
-};
-
-/**
- * Initialize EasyMDE editor
- *
- * @param {Element} element
- * @param {EasyMDE.Options} options
- * @return {EasyMDE}
- */
-export function createMarkdownEditor(element, options) {
-    const editor = new EasyMDE({
-        element,
-        ...defaultMarkdownEditorOptions,
-        ...options,
-    });
-
-    // overriding default CodeMirror shortcuts
-    editor.codemirror.addKeyMap({
-        Home: "goLineLeft", // move the cursor to the left side of the visual line it is on
-        End: "goLineRight", // move the cursor to the right side of the visual line it is on
-    });
-
-    // adding ability to fire events on the hidden element
-    if (element.dataset.listen) {
-        const events = element.dataset.listen.split(" ");
-        events.forEach((event) => {
-            try {
-                editor.codemirror.on(event, (e) => e.getTextArea().dispatchEvent(new Event(event)));
-            } catch (e) {
-                console.warn("Invalid event provided", event);
-            }
-        });
-    }
-
-    return editor;
-}
-
-export const imageUploadOptions = {
-    uploadUrl: imageUploadUrl,
-    uploadMethod: "POST",
-    uploadFieldName: "media",
-    jsonFieldName: "uploaded",
-    progressText: "![Загружаю файл...]()",
-    urlText: "![]({filename})",
-    errorText: "Ошибка при загрузке файла :(",
-    allowedTypes: [
-        "image/jpeg",
-        "image/png",
-        "image/jpg",
-        "image/gif",
-        "video/mp4",
-        "video/quicktime", // .mov (macOS' default record format)
-    ],
-    extraHeaders: {
-        Accept: "application/json",
-    },
-    extraParams: {
-        code: imageUploadCode,
-    },
-};
-
-export function handleFormSubmissionShortcuts(event) {
-    const isEnter = event.key === "Enter";
-    const isCtrlOrCmd = event.ctrlKey || event.metaKey;
-    const isEnterAndCmd = isEnter && isCtrlOrCmd;
-    if (!isEnterAndCmd) {
-        return;
-    }
-
-    const form = findParentForm(e.target);
-    if (!form || !isCommunicationForm(form)) {
-        return;
-    }
-
-    form.submit();
-}
 
 export function isMobile() {
     const userAgent = navigator.userAgent || navigator.vendor || window.opera;
